lib/survey: restore survey code from the deprecated https://github.com/Decawave/mynewt-dw1000-core.git repo
I assume while transition to a new repo from mynewt-dw1000-core
and probably because of the company acquision guys from Decawave
were in a hurry and commented lib/survey code out in order to
fix compilation errors. This leads to non-functional survey
code and outputting of not initialized memory. In this patch
I return original code, which seems to work fine.